Add Firmware files for 3.2.0
This release introduces the IMU support. This version was not intended for production as there was still a lot of experimenting with the core functionality for the esp32 WROOM DA module. The API has not been used correctly and this code will not compile until properly refactored. Therefore this will be marked as archive and not part of the series.

Code starting from version 3.2.0 is exclusive to ESP32, though it can be made suitable for Arduino AVR boards by making some adjustments.